separate class for unit tests with sling:alias, with having resource.resolver.optimize.alias.resolution disabled
